refactor: move `guessWorkerFormat()` to its own file
